+++ Unity.Entities.Graphics/DrawCommandGeneration.cs
@@ -117,11 +117,15 @@ namespace Unity.Rendering
         }
     }
 
+    [BurstCompile]
     internal unsafe struct ThreadLocalAllocator
     {
         public const int kInitialSize = 1024 * 1024;
         public const Allocator kAllocator = Allocator.Persistent;
         public static readonly int NumThreads = ChunkDrawCommandOutput.NumThreads;
+
+        private ProfilerMarker rewindAllocatorsMarker;
+        private ProfilerMarker rewindMarker;
 
         [StructLayout(LayoutKind.Explicit, Size = JobsUtility.CacheLineSize)]
         public unsafe struct PaddedAllocator
@@ -163,23 +167,26 @@ namespace Unity.Rendering
                 else
                     Allocators.ElementAt(i).Initialize(1);
             }
+
+            rewindAllocatorsMarker = new ProfilerMarker("RewindAllocators");
+            rewindMarker = new ProfilerMarker("Rewind");
         }
 
         public void Rewind()
         {
-            Profiler.BeginSample("RewindAllocators");
+            rewindAllocatorsMarker.Begin();
             for (int i = 0; i < NumThreads; ++i)
             {
                 ref var allocator = ref Allocators.ElementAt(i);
                 if (allocator.UsedSinceRewind)
                 {
-                    Profiler.BeginSample("Rewind");
+                    rewindMarker.Begin();
                     Allocators.ElementAt(i).Allocator.Allocator.Rewind();
-                    Profiler.EndSample();
+                    rewindMarker.End();
                 }
                 allocator.UsedSinceRewind = false;
             }
-            Profiler.EndSample();
+            rewindAllocatorsMarker.End();
 
         }