+++ CHANGELOG.md
@@ -4,6 +4,17 @@ uid: changelog
 
 # Changelog
 
+## [1.4.9-exp.2] - 2025-03-07
+
+### Changed
+
+* Optimized code in OnPerformCulling to use Burst compilation
+
+### Fixed
+
+* Fixed a rare null reference issue when using a motion vector pass when renderers had null materials
+
+
 ## [1.4.8] - 2025-02-24
 
 ### Fixed